src/python/requirements.txt
grpcio==1.47.0
grpcio-tools==1.47.0
python-decouple==3.6
